requirements.txt
asgiref
autopep8
django==3.1.7
pycodestyle
pytz
sqlparse
unipath==1.1
dj-database-url==0.5.0
python-decouple==3.4
gunicorn
whitenoise
requests==2.25.1
pytest==6.2.3
pytest-django==4.1.0
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12.4
psycopg2-binary
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0
django-filter==2.4.0
scikit-learn==1.1.0
numpy
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
django-crontab
pandas>1.1.0
pulp
holidays
matplotlib
pyomo

